requirements.txt
functions-framework
google-cloud-secret-manager
google-generativeai
gspread
google-auth-oauthlib
pytz
requests